+++ cocoa-beans-commands/src/main/java/net/apartium/cocoabeans/commands/parsers/CompoundParser.java
@@ -289,4 +289,9 @@ public class CompoundParser<T> extends ArgumentParser<T> implements GenericNode 
 
     }
 
+    @Override
+    public boolean isSupportMultipleArguments() {
+        return true;
+    }
+
 }